+++ test/src/Ouzo/Core/ValidatableTest.php
@@ -263,7 +263,7 @@ class ValidatableTest extends PHPUnit_Framework_TestCase
     /**
      * @test
      */
-    public function shouldValidateIsEmptyObject()
+    public function notEmptyShouldReturnErrorForEmpty()
     {
         //given
         $object = '';
@@ -279,7 +279,7 @@ class ValidatableTest extends PHPUnit_Framework_TestCase
     /**
      * @test
      */
-    public function shouldNotValidateWhenPassObject()
+    public function notEmptyShouldNotReturnErrorForNotEmpty()
     {
         //given
         $object = new stdClass();
@@ -287,6 +287,38 @@ class ValidatableTest extends PHPUnit_Framework_TestCase
 
         //when
         $validatable->validateNotEmpty($object, 'error1');
+
+        //then
+        $this->assertEmpty($validatable->getErrors());
+    }
+
+    /**
+     * @test
+     */
+    public function emptyShouldReturnErrorForNonEmpty()
+    {
+        //given
+        $object = new stdClass();
+        $validatable = new Validatable();
+
+        //when
+        $validatable->validateEmpty($object, 'error1');
+
+        //then
+        Assert::thatArray($validatable->getErrors())->containsOnly('error1');
+    }
+
+    /**
+     * @test
+     */
+    public function emptyShouldNotReturnErrorForEmpty()
+    {
+        //given
+        $object = '';
+        $validatable = new Validatable();
+
+        //when
+        $validatable->validateEmpty($object, 'error1');
 
         //then
         $this->assertEmpty($validatable->getErrors());