+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Enhancements:
 * [ORM] Added possibility of using multiple Restrictions in Any::of for the same key.
 * [Utilities] Added Suppliers::memoizeWithExpiration that returns supplier which caches the result of function.
 * [Core] Minimal log level can be assigned to a particular class/name in logger configuration.
+* [Core] Added Validatable::validateEmpty method.
 
 Bug fixes:
 * [Localization] Fixed I18n::loadLabels not to load translation file if it was already loaded (issue #173).